add support for "ignored" plugins
when calling `handleRequires()`, it's helpful for the worker to _not_ load certain plugins.  if all are loaded, it would need to explicitly omit certain finalized plugins from the options sent to the `Mocha` constructor.  this adds an `ignoredPlugins` option to `handleRequires()`, which passes thru to the `PluginLoader` constructor.  it says "even though we have definitions for plugins _x_, _y_, and _z_, do not load them from `--require`'d modules".  this will save us some cycles since we now filter out those _up-front_ instead of after the fact.

- also add `unexpected-set` and `unexpected-map` for better assertions when working with `Set` and `Map` objects (these do not export globals)
- adds a custom
- refactor `PluginLoader` constructor (and `PluginLoader.create()`) to accept a single options object instead (of type `PluginLoaderOptions`)